replace old migration with new one.
I was running into this weird issue where my first migration wasn't
connecting to the rails console or the database. So I couldn't call
Thought.new in the console or in code. I decided to regenerate the
migration that creates the thoughts table. Seems to have worked.
Although i'm not totally sure why