add in rails12factor for heroku.
Heroku was needed to be set up eariler than originally thought. It looks
like the Reddit API requires a redirect url to be set up. It also
requires that I use OAuth2 to log in as my user. This is a bigger
problem than originally anticipated. I will continue to work with this
project and get the oauth problem solved.

I need to figure out whether I want to use the reddit-omniauth gem,
which would work. or if I should just hand roll this bad boy. as of
right now I am leaning toward the gem.